Relocate the XML based metadata source models
